feat: make geometry a Cython extension
Current code requires very sequential code to be run in order to obtain
the crossing of the polygon points. In order to speed it up without
creating extra runtime dependencies to the end user, we use cython to
write it. This only makes the building process slightly more annoying
but should be ok.